requirements.txt
streamlit
streamlit-ace
streamlit-autorefresh
huggingface-hub
python-dotenv
requests
firecrawl-py